Moved correct_mj struct to private header file
